server/audio_service/requirements.txt
Flask
gTTS
PyMuPDF
python-dotenv
requests